Added new test, and removed stub_omniauth to rails_helper
