+++ spec/features/messaging/user_can_send_message_spec.rb
@@ -1,0 +1,7 @@
+require 'rails_helper'
+
+RSpec.feature "user can send a message to another user" do
+  describe "user sees mailbox link in navbar" do
+    
+  end
+end
