+++ spec/features/user/user_can_log_in_with_google_spec.rb
@@ -14,21 +14,3 @@ RSpec.feature "user logs in" do
     expect(page).to have_content("Logout")
   end
 end
-
-def stub_omniauth
-  OmniAuth.config.test_mode = true
-  OmniAuth.config.mock_auth[:google] = OmniAuth::AuthHash.new({
-    provider: "google_oauth2",
-      uid: "12345678910",
-      info: {
-        email: "limsammy1@gmail.com",
-        first_name: "Sam",
-        last_name: "Lim"
-      },
-      credentials: {
-        token: "abcdefg12345",
-        refresh_token: "abcdefg12345",
-        expires_at: DateTime.now,
-      }
-    })
-end
