+++ spec/rails_helper.rb
@@ -58,3 +58,21 @@ RSpec.configure do |config|
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
 end
+
+def stub_omniauth
+  OmniAuth.config.test_mode = true
+  OmniAuth.config.mock_auth[:google] = OmniAuth::AuthHash.new({
+    provider: "google_oauth2",
+      uid: "12345678910",
+      info: {
+        email: "limsammy1@gmail.com",
+        first_name: "Sam",
+        last_name: "Lim"
+      },
+      credentials: {
+        token: "abcdefg12345",
+        refresh_token: "abcdefg12345",
+        expires_at: DateTime.now,
+      }
+    })
+end
